# perf: faster hash marshaling methods