@cristaloleg 2024-05-27
@renaynay added a fuzzing test for `hexToUpper` (to run: `go test -v -run=^$ -fuzz=Fuzz_hexToUpper .`)